requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.7
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.9
Markdown==3.7
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.7
pre_commit==4.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.29.3
